+++ local-s3-rest/src/main/java/com/robothy/s3/rest/handler/GetObjectController.java
@@ -82,6 +82,7 @@ class GetObjectController implements HttpRequestHandler {
 
       getObjectAns.getUserMetadata().forEach((k, v) -> response.putHeader(AmzHeaderNames.X_AMZ_META_PREFIX + k, v));
       ObjectLockHeaders.addHeaders(response, getObjectAns.getObjectLock());
+      RestoreObjectController.addRestoreHeader(response, getObjectAns.getRestoreExpiryDate());
       CustomerEncryptionHeaders.addHeaders(response, getObjectAns.getCustomerEncryption());
       ServerSideEncryptionHeaders.addHeaders(response, getObjectAns.getServerSideEncryption(), false);
       if (ChecksumHeaders.isChecksumModeEnabled(request)) {