+++ local-s3-rest/src/main/java/com/robothy/s3/rest/handler/HeadObjectController.java
@@ -77,6 +77,7 @@ class HeadObjectController implements HttpRequestHandler {
       ResponseUtils.addETag(response, object.getEtag());
       object.getUserMetadata().forEach((k, v) -> response.putHeader(AmzHeaderNames.X_AMZ_META_PREFIX + k, v));
       ObjectLockHeaders.addHeaders(response, object.getObjectLock());
+      RestoreObjectController.addRestoreHeader(response, object.getRestoreExpiryDate());
       CustomerEncryptionHeaders.addHeaders(response, object.getCustomerEncryption());
       ServerSideEncryptionHeaders.addHeaders(response, object.getServerSideEncryption(), false);
       if (ChecksumHeaders.isChecksumModeEnabled(request)) {